Now using color curves similar to BGB.
After playing for a while with the color settings of BGB and color pattern ROMs, it looks like the color curves are essentially a smoothstep of each compoenent with the final green component being a blend of the 3 input components, this followed by a bit of desaturation and gamma correction.
The information is precomputed to reduce the cost of handling ROMs with Hicolor images.